requirements.txt
# Зависимости дипломного backend-проекта (Django + DRF + фоновые задачи)
django~=5.0
djangorestframework~=3.15.0
celery~=5.3.0
redis~=5.0.0
requests~=2.31.0
ujson~=5.9.0
pyyaml~=6.0.0
django-rest-passwordreset>=1.3.0
python-dotenv~=1.0.0
pytest~=8.0.0
pytest-django~=4.8.0
